Added docs and cleanup
