+++ graphql/src/commonMain/kotlin/com/suparnatural/core/graphql/GraphQlOperation.kt
@@ -5,11 +5,10 @@ import kotlinx.serialization.KSerializer
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.UnstableDefault
 import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.JsonElement
 import kotlinx.serialization.json.JsonObject
 
 /**
- * Represents a GraphQl response error.
+ * A serializable GraphQl error as per the spec
  */
 @Serializable
 data class GraphQlResponseError(
@@ -22,7 +21,7 @@ data class GraphQlResponseError(
 }
 
 /**
- * Represents GraphQl response
+ * A serializable GraphQl response as per the spec
  */
 @Serializable
 data class GraphQlResponse<T>(
@@ -31,7 +30,9 @@ data class GraphQlResponse<T>(
         val extensions: JsonObject? = null
 )
 
-// Serializable GraphQlRequest
+/**
+ * A serializable GraphQl request as per the spec
+ */
 @Serializable
 data class GraphQlRequest(
         val operationName: String,
@@ -40,18 +41,43 @@ data class GraphQlRequest(
 )
 
 /**
- * A GraphQl operation consists of a request and the corresponding response.
+ * A GraphQl operation encapsulates the query / mutation
+ * and its corresponding response definition.
  */
 abstract class GraphQlOperation<T> {
+    /**
+     * complete GraphQl query serialized as a string
+     */
     abstract val source: String
+
+    /**
+     * name of the operation. It is sent as `operationName`
+     * in the request.
+     */
     abstract val name: String
+
+    /**
+     * A map of operation variables
+     */
     abstract val variables: MutableMap<String, Any?>
+
+    /**
+     * A mutable map which holds context of the operation
+     * which can be used by links down the chain
+     */
     val context: MutableMap<String, Any> = mutableMapOf()
 
+    /**
+     * A serialized json string of the GraphQl operation
+     */
     @UnstableDefault
-    val serializedString: String
+    val jsonString: String
         get() {
             return Json.stringify(GraphQlRequest.serializer(), GraphQlRequest(name, source, variables))
         }
+
+    /**
+     * Serializer to parse the response body
+     */
     abstract val responseSerializer: KSerializer<T>
 }
